Remove PTRACE from capabilities
This was added when I was trying to figure out why clamdscan wasn't
connecting to the local clamd on 127.0.0.1:3310, and is not needed in a
production situation.  It *did* allow me to `strace clamdscan ...` and
find out that not specifying the `TCPAddr` field (as "0.0.0.0") in the
clamd.conf configuration file was causing the auto-detection logic to
try IPv6 and fail, halting the scan with an unprinted error.